# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ECONEURA API",
    "description": "Enterprise AI Ecosystem - Backend API Documentation",
    "version": "3.0.0",
    "contact": {
      "name": "ECONEURA Support",
      "email": "support@econeura.com",
      "url": "https://econeura.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://econeura.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Development server"
    },
    {
      "url": "https://econeura-backend-prod.azurewebsites.net/api",
      "description": "Production server (Azure)"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new user",
        "description": "Create a new user account with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string", "minLength": 6},
                  "name": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AuthResponse"}
              }
            }
          },
          "400": {"description": "Invalid input"},
          "409": {"description": "Email already exists"}
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "description": "Authenticate user with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AuthResponse"}
              }
            }
          },
          "401": {"description": "Invalid credentials"}
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "description": "Get new access token using refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AuthResponse"}
              }
            }
          },
          "401": {"description": "Invalid refresh token"}
        }
      }
    },
    "/auth/user": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user",
        "description": "Get authenticated user information",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "User information",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/User"}
              }
            }
          },
          "401": {"description": "Unauthorized"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "name": {"type": "string"},
          "role": {"type": "string", "enum": ["user", "admin"]}
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {"type": "string"},
          "refreshToken": {"type": "string"},
          "expiresIn": {"type": "integer"},
          "user": {"$ref": "#/components/schemas/User"}
        }
      },
...
    }
  }
}